[fix] Fix the issue of 'Unable to find any supported Python versions' in the deployment of vercel
