+++ package.json
@@ -1,0 +1,5 @@
+{
+	"engines": {
+		"node": "18.x"
+	}
+}
